separate out wasm from client lib
- revise client lib for easier upgrading through usage of wasm-pack
  with --omit-default-module-path
- add tss-lib-wasm package that separately holds the wasm file. this
  allows loading the wasm from a package on demand and keeping the
  size of tss-lib low.
