+++ packages/tss-lib-wasm/.eslintrc.json
@@ -1,0 +1,6 @@
+{
+  "extends": ["../../.eslintrc"],
+  "parserOptions": {
+    "project": "./tsconfig.json"
+  }
+}
