+++ packages/tss-lib-wasm/.gitignore
@@ -1,0 +1,2 @@
+# typescript
+/dist/
