+++ packages/tss-lib-wasm/src/dkls_bg.wasm.d.ts
