+++ packages/tss-lib-wasm/src/index.ts
@@ -1,0 +1,3 @@
+import wasm from "./dkls_bg.wasm";
+
+export default { wasm, keyType: "secp256k1" };
