+++ packages/tss-lib-wasm/torus.config.js
@@ -1,3 +1,4 @@
 module.exports = {
   ...require("@toruslabs/config/torus.config"),
+  esm: false,
 };
