+++ packages/tss-lib-wasm/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../tsconfig.json",
+  "exclude": ["*.d.ts"],
+  "include": ["./src/**/*.ts"]
+}
