+++ packages/tss-lib-wasm/webpack.config.js
@@ -1,0 +1,27 @@
+
+
+module.exports = {
+  umdConfig: {
+    module: {
+      rules: [
+        {
+          // Pack Wasm inline.
+          test: /\.wasm$/,
+          type: "asset/inline",
+        },
+      ],
+    }
+  },
+  cjsConfig: {
+    module: {
+      rules: [
+        {
+          // Pack Wasm inline.
+          test: /\.wasm$/,
+          type: "asset/inline",
+        },
+      ],
+    }
+  }
+};
+
