--- packages/tss-lib/babel.config.js
@@ -1,1 +1,0 @@
-module.exports = require("@toruslabs/config/babel.config");
