--- packages/tss-lib/rollup.config.mjs
@@ -1,3 +1,0 @@
-export default {
-  input: "./browser.js",
-};
