--- packages/tss-lib/tsconfig.build.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["./"]
-}
