--- packages/tss-lib/webpack.config.js
@@ -1,3 +1,0 @@
-exports.baseConfig = {
-  entry: "./browser.js",
-};
